+++ src/js/audio/util.test.ts
@@ -1,34 +1,9 @@
 import { describe, test, expect } from "@jest/globals";
+import {
+  MockSpeechRecognitionResult,
+  MockSpeechRecognitionResultList,
+} from "./mocks";
 import { getCommand } from "./util";
-
-// Web Speech API classes aren't available in Node, so create mocks
-
-class SpeechList<Type> {
-  [index: number]: Type;
-  length: number;
-
-  constructor(items: Type[]) {
-    this.length = items.length;
-    for (const [index, item] of items.entries()) {
-      this[index] = item;
-    }
-  }
-
-  item(index: number) {
-    return this[index];
-  }
-}
-
-class MockSpeechRecognitionResult extends SpeechList<SpeechRecognitionAlternative> {
-  isFinal: boolean;
-
-  constructor(items: SpeechRecognitionAlternative[], isFinal: boolean) {
-    super(items);
-    this.isFinal = isFinal;
-  }
-}
-
-class MockSpeechRecognitionResultList extends SpeechList<SpeechRecognitionResult> {}
 
 describe("getting command", () => {
   test("single word", () => {
